Introduce dirty hack in CVE-2023-24042 config
Previously, server kept returning its docker-specific IP
for data connections. The ftp command reported IP mismatch
when connecting to localhost, and finding out the docker IP
to connect to seems to be too tedious.

By setting the mask to all 1s it forces the server to always
return the external_ip for data connections.